fix todo client
